requirements.txt
arangodb
alembic==1.0.0
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.7.0
certifi==2018.8.13
cffi==1.11.5
chardet==3.0.4
click==6.7
constantly==15.1.0
cryptography==2.3.1
cssselect==1.0.3
dominate==2.3.1
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Migrate==2.2.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
future==0.16.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
itsdangerous==0.24
Jinja2==2.10
lxml==4.2.4
Mako==1.0.7
MarkupSafe==1.0
parsel==1.5.0
pbr==4.2.0
psycopg2==2.7.5
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
python-arango==4.2.1
python-dateutil==2.7.3
python-editor==1.0.3
queuelib==1.5.0
requests==2.19.1
Scrapy==1.5.0
service-identity==17.0.0
six==1.11.0
SQLAlchemy==1.2.11
stevedore==1.29.0
Twisted==18.7.0
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
visitor==0.1.3
w3lib==1.19.0
Werkzeug==0.14.1
WTForms==2.2.1
zope.interface==4.5.0
